Move Hikari stuff into its own package
