+++ app/src/main/java/gov/hhs/cdc/trustedintermediary/external/javalin/App.java
@@ -6,11 +6,11 @@ import gov.hhs.cdc.trustedintermediary.domainconnector.DomainConnector;
 import gov.hhs.cdc.trustedintermediary.domainconnector.DomainConnectorConstructionException;
 import gov.hhs.cdc.trustedintermediary.domainconnector.DomainResponseHelper;
 import gov.hhs.cdc.trustedintermediary.domainconnector.UnableToReadOpenApiSpecificationException;
-import gov.hhs.cdc.trustedintermediary.external.HikariConnectionPool;
 import gov.hhs.cdc.trustedintermediary.external.apache.ApacheClient;
 import gov.hhs.cdc.trustedintermediary.external.azure.AzureDatabaseCredentialsProvider;
 import gov.hhs.cdc.trustedintermediary.external.azure.AzureSecrets;
 import gov.hhs.cdc.trustedintermediary.external.hapi.HapiFhirImplementation;
+import gov.hhs.cdc.trustedintermediary.external.hikari.HikariConnectionPool;
 import gov.hhs.cdc.trustedintermediary.external.inmemory.KeyCache;
 import gov.hhs.cdc.trustedintermediary.external.inmemory.LoggingMetricMetadata;
 import gov.hhs.cdc.trustedintermediary.external.jackson.Jackson;
